[Analyzer] Move SpecsState into an own file
